+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 <!-- markdownlint-disable MD024 -->
 # Changelog
 
-## [Unreleased]
+## [1.1.3] - 2025-10-28
 
 ### Added